Code_Source/interface/requirements.txt
flask>=2.0.1
gunicorn>=20.1.0
kafka-python>=2.0.2